refactor: standardize file naming conventions
- Rename controllers to use .controller.js suffix
- Rename middleware to use .middleware.js suffix
- Rename routes to use .route.js suffix
- Add cohorts model structure
- Update imports in server.js to match new naming
